feat: create stardust component interfaces index
