+++ client/src/app/components/stardust/interfaces/IPreExpandedConfig.tsx
@@ -1,0 +1,10 @@
+export interface IPreExpandedConfig {
+    isPreExpanded?: boolean;
+    unlockConditions?: boolean[];
+    features?: boolean[];
+    immutableFeatures?: boolean[];
+    nativeTokens?: boolean[];
+
+    // generic to expand all
+    isAllPreExpanded?: boolean;
+}
