+++ client/src/app/components/stardust/interfaces/index.tsx
@@ -1,10 +1,1 @@
-export interface IPreExpandedConfig {
-    isPreExpanded?: boolean;
-    unlockConditions?: boolean[];
-    features?: boolean[];
-    immutableFeatures?: boolean[];
-    nativeTokens?: boolean[];
-
-    // generic to expand all
-    isAllPreExpanded?: boolean;
-}
+export * from './IPreExpandedConfig';
